Edited to run with heroku
Specified Heroku should use the Ahead Of Time (AOT) compiler to create the dist folder (where the javascript and html converted version of the app will be run.)
Installed Express server to run the app from dist folder.
Added Node and NPM engines for Heroku to use to run the applicaion.
Changed:
-Typescipt
-@angular.cli
-@angular/compiler/cli
from Devdependencies to Dependencies in package.json

Source for information on converting to work with Heroku: https://itnext.io/how-to-deploy-angular-application-to-heroku-1d56e09c5147